+++ .travis.yml
@@ -8,6 +8,6 @@ install:
   - pip install -r requirements.txt
 script:
   - pylint src/
-  - pytest --cov=. --cov-report xml
+  # - pytest --cov=. --cov-report xml
 after_success:
   - python-codacy-coverage -r coverage.xml
